Adicionando funcionalidades de integração.

+++ server/SportPoint.Server.Integration/IGenericIntegration.cs
@@ -61,6 +61,6 @@ namespace SportPoint.Server.Integration
         /// Método que retorna a quantidade de registros da entidade.
         /// </summary>
         /// <returns>Quantidade de registros encontrados.</returns>
-        int Count();
+        long Count();
     }
 }
